Move code into subfolders.
