+++ app/src/main/AndroidManifest.xml
@@ -31,11 +31,11 @@
             android:value="@string/BLOOP_GOOGLE_MAPS_API_KEY" />
 
         <activity
-            android:name=".TutorialActivity"
+            android:name=".activities.TutorialActivity"
             android:screenOrientation="portrait" />
 
         <activity
-            android:name=".PlayLoginActivity"
+            android:name=".activities.PlayLoginActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -44,11 +44,11 @@
         </activity>
 
         <activity
-            android:name=".BloopActivity"
+            android:name=".activities.BloopActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".FlagCreationActivity" />
+        <activity android:name=".activities.FlagCreationActivity" />
     </application>
 
 </manifest>
