+++ app/src/main/java/website/bloop/app/sound/BloopSoundPlayer.java
@@ -1,10 +1,12 @@
-package website.bloop.app;
+package website.bloop.app.sound;
 
 import android.content.Context;
 import android.media.AudioAttributes;
 import android.media.AudioManager;
 import android.media.SoundPool;
 import android.os.Build;
+
+import website.bloop.app.R;
 
 public class BloopSoundPlayer {
     private final Context mContext;
