+++ app/src/main/java/website/bloop/app/views/FlagView.java
@@ -1,4 +1,4 @@
-package website.bloop.app;
+package website.bloop.app.views;
 
 import android.content.Context;
 import android.graphics.Canvas;
