+++ app/src/main/res/layout/activity_flag_creation.xml
@@ -8,14 +8,14 @@
     <view
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        class="website.bloop.app.BackgroundLandscapeView"
+        class="website.bloop.app.views.BackgroundLandscapeView"
         android:id="@+id/background_view" />
 
     <RelativeLayout
         android:id="@+id/ui_flag_creation"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="website.bloop.app.FlagCreationActivity"
+        tools:context="website.bloop.app.activities.FlagCreationActivity"
         android:layout_marginBottom="@dimen/activity_vertical_margin"
         android:layout_marginEnd="@dimen/activity_horizontal_margin"
         android:layout_marginStart="@dimen/activity_horizontal_margin"
@@ -36,7 +36,7 @@
             <view
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
-                class="website.bloop.app.FlagView"
+                class="website.bloop.app.views.FlagView"
                 android:id="@+id/flag_view"
                 android:layout_weight="4" />
         </LinearLayout>
