+++ app/src/main/res/layout/activity_play_login.xml
@@ -5,12 +5,12 @@
     android:id="@+id/activity_play_login"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="website.bloop.app.PlayLoginActivity">
+    tools:context="website.bloop.app.activities.PlayLoginActivity">
 
     <view
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        class="website.bloop.app.SonarView"
+        class="website.bloop.app.views.SonarView"
         android:layout_centerHorizontal="true"
         android:layout_centerVertical="true"
         android:id="@+id/sonarView" />
